Add API for HttpClient (#11)
Add API for HttpClient which can be used to process HTTP requests. The concrete implementation of this client must be implemented by the framework specific Silhouette binding. The client is no full-blown HTTP client implementation. It implements only the parts which Silhouette depends on, and it is only meant for internal usage.