Remove session based functionality
A session is a non-standardized concept that is mostly implemented by a framework or the standard lib of a programming language. Play uses a stateless approach by using a cookie to store session data on the client. There exists also a stateful approach which stores session data on the server side and a reference (ID) to this data on the client side. Silhouette already supports this concepts with cookies and stateful or stateless authenticators. There is also no necessity to support a custom session approach like Play does. This would bring only complexity to the Silhouette API.